Migrate training to class
Moving the training steps to a standalone class makes it easier to store
and update models, optimizers, training sets, and other necessary
components for the training into a single instance of the class. This
also will help differentiate new models that are added to the tool down
the road by giving each model its own class to handle training.

Signed-Off-By: Robert Clark <robdclark@outlook.com>