Increase dataset workers to eliminate bottleneck
The dataset streams in multiple batches of images and performs various
pre-processing tasks on each image which is both disk and CPU-intensive
as the images are loaded from storage and transformations are made. This
process can be improved by adding an increased number of parallel
workers for the training data loader to queue up multiple pipelines
which independently load images from the disk and perform operations
before sending it to the main process to add to the dataset, allowing
data to be sent to the GPUs (or CPUs for non-GPU systems) faster for a
more optimized pipeline.

Signed-Off-By: Robert Clark <robdclark@outlook.com>