Support distributed training
Using the DistributedDataParallel and DistributedSampler modules, the
application should support multi-GPU workloads by dividing the training
and testing datasets between all available GPUs. To allow the
application to be launched as simply as possible, it leverages the
multiprocessing.spawn functionality built-in to PyTorch to launch
individual processes on each GPU while allowing the GPUs to communicate
over NCCL during the training process.

By default, the application should leverage all available GPUs on a
system to maximize training throughput. To limit the number of GPUs used
or to force CPU-only mode, the --gpus flag can be used for subsets of
GPUs.

Signed-Off-By: Robert Clark <robdclark@outlook.com>