Add parameter to specify data workers
Increasing the number of workers for the data loaders can improve the
performance of the application by minimizing the amount of time spent on
reading and preprocessing images, causing the compute devices to be
starved for data. Creating a CLI argument makes it easy to experiment
with the proper value to optimize performance.

Signed-Off-By: Robert Clark <robdclark@outlook.com>