Save best and latest models
Given the relative unreliability of using PSNR to evaluate images, the
model with the highest PSNR might not necessarily have the highest
perceptual accuracy. With this in mind, saving both the "best" model and
the latest model allows for comparisons between the two where the later
model could very well be better than the latest.

Signed-Off-By: Robert Clark <robdclark@outlook.com>