Add distributed support with torchrun
Torchrun allows for multi-GPU and multi-node training with the ability
to do elastic computing. Torchrun takes several parameters and
automatically creates environment variables which can be parsed from the
application.

In order to support torchrun, the --gpus parameter has been removed.
This changes the default behavior of running on all available GPUs on a
system and instead running on a single-GPU if not otherwise specified.
While the ultimate goal is to be able to run on multi-GPU by default,
that can be added down the road as a wrapper for torchrun which selects
the number of GPUs based on the head node.

Signed-Off-By: Robert Clark <robdclark@outlook.com>