Create a base class for trainers
A large chunk of the trainer classes for the SRGAN and ESRGAN models is
identical. Moving the identical code to a base class that can be
inherited in the specific trainer classes makes it easier to maintain
changes between the models and dramatically reduces the amount of code
in the repository.

Signed-Off-By: Robert Clark <robdclark@outlook.com>