Add dataset multiplication to test set
The testing dataset was not being multiplied like the training dataset,
causing validations to be shorter than expected. Additionally, the
throughput calculations were off as it assumed the dataset was the
multiplied size when it was far less.

Multiplying the test dataset will allow for a greater range of images
to validate against to more properly view validation loss and overall
model performance.

Signed-Off-By: Robert Clark <robdclark@outlook.com>